applications/analytics_application/analytics_application/__init__.py
from configurations import importer
import os

importer.install()

os.environ.setdefault("DJANGO_SETTINGS_MODULE", "analytics_application.settings")
os.environ.setdefault("DJANGO_CONFIGURATION", "Dev")
